requirements.txt
python-telegram-bot>=20.0
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
gunicorn>=20.1.0
eventlet>=0.30.0
psycopg2-binary>=2.9.0  # 如果使用PostgreSQL
pymysql>=1.0.0  # 如果使用MySQL 